moved reamining functions to eval_tools.py
